+++ changelog.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog][keep a changelog] and this project adh
 - Implemented 3D secure for payment cards requiring authentication.
 - Add missing locale to pelcro's native buttons [PRIVATE]
 
+### Fixed
+
+- Only skip card if 100% coupon code is set to forever
+
 ## [Released]
 
 ## [0.12.1] - 2021-04-26